extract the utils duplicated between tests and web
