+++ tests/test_grammar_validator.py
@@ -3,8 +3,7 @@ import unittest
 from pegen.grammar import Grammar
 from pegen.grammar_parser import GeneratedParser as GrammarParser
 from pegen.validator import SubRuleValidator, ValidationError
-
-from tests.utils import parse_string
+from pegen.utils import parse_string
 
 
 class TestPegen(unittest.TestCase):
