+++ tests/test_grammar_visualizer.py
@@ -3,7 +3,7 @@ from typing import List
 
 from pegen.grammar_parser import GeneratedParser as GrammarParser
 from pegen.grammar_visualizer import ASTGrammarPrinter
-from tests.utils import parse_string
+from pegen.utils import parse_string
 
 
 def test_simple_rule() -> None:
